errors: Add an exception handling module
As it name implies, the `errors.py` module will be used to define all
the exception-specific functionality of the haproxy module.

At the moment it just includes the `HAProxyGenericError`, the base
exception for the current module, and the `CommandExecError` which
denotes that a failure occurred during command execution. Since a
command's output is intended to be returned as a list of strings, the
`HAProxyGenericError` checks whether the exception message is a list and
if so concatenates it into a single string.