utils: Introduce a module to define utility functions
This commit introduces a dedicated module to define utility functions.

Currently, we add the `cmd_succeeded` function which will be used to
check if a given command's output contains an error. In HAProxy's
Runtime API, most commonly a succeeded command returns an empty string.
However, some commands even though have succeeded return an informative
message to the caller. This helper function checks the status of a given
output and if it's empty or the message is included in the
`ignored_outputs` argument, it marks the command status as successful.